+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `as_str` filter
 - Way fewer allocations and significant speedup (2-5x) for templates with large objects/loops
+- Checks that all macro files are accounted for at compile time and errors if it's not the case
 
 ## 0.11.12 (2018-08-04)